Get and store audio file urls to href_list
There are a number of cases we have to handle:

0. No match. We get a result summary page telling us we have no matches.
1. Only one match. We get the dictionary page for the word.
2. More than one match. We get a result summary page with links to dictionary pages

For case 2 the results are divided among match on main word and match on inflected form.
Probably best to drop the inflected forms if there are matches on the main word.
This has been implemented by getting the first ul tag. If there are exact
matches, they will be in the first ul. If there are no main word matches,
the first ul will be the list of matching inflections.

The table with id="ord" is only available on word pages so I use this as a
test. Could have used audio tag as well I guess.

I see no point in downloading both ogg and mp3 but perhaps it would be
better to let a parameter decide which one is prefered. Thus, power users
could change their preference easier.