Add code for search and grab results
The search results are found in the first div with
class="searchResultBox". This div is present even if no matches were
found. We can get all the links and use their href attributes for further
processing. If no results, the list of links will be empty so we don't
have to fetch the icon.