Add code for making request for a word to self.url
Lexin uses GWT-RPC for communicating with the server. The response is
parsed and presented using javascript. The request structured was
uncovered by inspecting the requests sent by the browser. The minimum
headers were identified using trial and error. The payload could
potentially be shortened if anyone feels inclined to read the gwt docs.