Add parsing and downloading of results
The response is almost valid json. It starts with '//OK' which can just be
dropped. It contains hex escapes which are invalid in json. These need to
be replaced with unicode escapes. This is done using
`.replace('\\x', '\\00')`. After this we have valid json that can be
parsed by the standard json module.

The json encodes a list and it turns
out that this list should be read from the end. The third last element
contains all the interesting information. This element contains yet
another list whose elements 3:-2 are strings containing xml that defines
all the interesting data about each word matching the search.

The xml is then parsed using BeautifulSoup. The url for the audio file is
the file attribute of the `<phonetic>` tag.