Add spelling update and part of speech extraction
The part of speech (Swedish abbreviation) is contained in the `type`
attribute of the `<lemma>` tag.

The search function matches any form, of the word, so we need to make sure
that we display the word in the form that is in the recording. The
spelling of the word is contained in the `value` attribute of the
`<lemma>` tag.