Avoid double downloads
Some v1 files are still used. These are easy to distinguish thanks to
their typical file names. The flag `v1_downloaded` is introduced to
keep track of whether a file `<word>.mp3` has been downloaded. If so,
there is no need to download again. We also perform the v1_download if
the search request fails.